[feat] 상품커스텀리파지토리 클래스추가

+++ boiler/src/main/java/com/bronze/boiler/util/Response.java
@@ -11,8 +11,8 @@ import java.util.List;
 @Getter
 public class Response<T> {
 
-    private int total;
-    private int currentPage;
+    private Long total;
+    private Long currentPage;
 
     private List<T> list;
